Make utilities module for functions like poly_eval
Add basic documentation to utils module; for now, pub use
poly_eval at the crate root.
